requirements.txt
Gunicorn
Flask
Pillow
opencv-python
numpy
scipy
Flask-CORS